% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{data}
\name{location}
\alias{location}
\title{Location information of the water pumps}
\format{
A tibble with 299 rows and 6 variables

\describe{
\item{pumpid}{ID number of the water pump}
\item{description}{Name of the pump location}
\item{x_arc1960}{Easting in metres, Arc 1960 / UTM zone 37S}
\item{y_arc1960}{Northing in metres, Arc 1960 / UTM zone 37S}
\item{lat_wgs84}{Latitude in WGS 84 coordinate system}
\item{long_wgs84}{Longitude in WGS 84 coordinate system}
}
}
\source{
British Geological Survey, National Geoscience Data Centre: \url{https://metadata.bgs.ac.uk/geonetwork/srv/eng/catalog.search#/metadata/9ed6c5d2-124b-50ed-e054-002128a47908}
}
\usage{
location
}
\description{
Location information of the water pumps
}
\keyword{datasets}
