% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyvol2015.R
\docType{data}
\name{weeklyvol2015}
\alias{weeklyvol2015}
\title{Weekly volume in year 2015}
\format{
A tibble with 297 rows and 53 variables

\describe{
\item{pumpid}{ID number of the water pump}
\item{1, 2, ..., 52}{Volume of water pumped in week 1 to 52 of 2015, one column per week}
}
}
\source{
British Geological Survey, National Geoscience Data Centre: \url{https://metadata.bgs.ac.uk/geonetwork/srv/eng/catalog.search#/metadata/9ed6c5d2-124b-50ed-e054-002128a47908}
}
\usage{
weeklyvol2015
}
\description{
Weekly volume in year 2015
}
\keyword{datasets}
