% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyvol2014.R
\docType{data}
\name{weeklyvol2014}
\alias{weeklyvol2014}
\title{Weekly volume in year 2014}
\format{
A tibble with 324 rows and 53 variables

\describe{
\item{PumpID}{ID number of the water pump}
\item{n}{Each column indicates the week number 1, 2, 3, ..., 52}
}
}
\usage{
weeklyvol2014
}
\description{
Weekly volume in year 2014
}
\keyword{datasets}
