% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyvol2015.R
\docType{data}
\name{weeklyvol2015}
\alias{weeklyvol2015}
\title{Weekly volume in year 2015}
\format{
A tibble with 297 rows and 53 variables
\describe{
\item{PumpID}{ID number of the water pump}
\item{n}{Each column indicates the week number 1, 2, 3, ..., 52}
}
}
\usage{
weeklyvol2015
}
\description{
Weekly volume in year 2015
}
\keyword{datasets}
